requirements.txt
flask
werkzeug
google-cloud-aiplatform